+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Added option to have hidden workshops
 - Improved SEO
 - Added workshop sign in kiosk
+- oved to tales-from-a-dev/tailwind-merge-phpMoved to tales-from-a-dev/tailwind-merge-php
 
 ## 1.3.5